// src/app/core/main/tasks/interfaces/task.interface.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface ITask {
    date:          string;
    type:          string;
    estimatedTime: string;
    delegationId:  number;
    comment:       Comment;
}

export interface Comment {
    content: string;
}

// Generated by https://quicktype.io

export interface ITaskGet {
    id:            number;
    date:          Date;
    state:         string;
    type:          string;
    estimatedTime: Date;
    delegationId:  number;
    comments:      CommentGet[];
    createdAt:     Date;
    updateAt:      Date;
    delegation:    Delegation;
}

export interface CommentGet {
    id:        number;
    content:   string;
    createdAt: Date;
    updateAt:  Date;
    taskId:    number;
}

export interface Delegation {
    id:       number;
    employee: Employee;
    consumer: Consumer;
}

export interface Consumer {
    id:         number;
    isCustomer: boolean;
    type:       string;
    person:     Person;
}

export interface Person {
    dni:       string;
    firstName: string;
    lastName:  string;
}

export interface Employee {
    id:       number;
    username: string;
    role:     string;
    person:   Person;
}